step1 refactoring
