+++ src/main/java/javablackjack/blackjack/io/Output.java
@@ -1,9 +1,8 @@
 package javablackjack.blackjack.io;
 
-import javablackjack.blackjack.GameResult;
+import javablackjack.blackjack.domain.GameResult;
 
 public class Output {
-
 
     public static void resultPrint(GameResult gameResult) {
         System.out.println(gameResult.getGameResult());
